Module jet_margin_pool::cpi

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

state

Structs

Return

Functions

close_loan

collect

configure

create_pool

deposit

margin_borrow

margin_refresh_position

margin_repay

register_loan

repay

withdraw