Module jet_margin_pool::instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Modules

state

Instruction struct definitions for #[state] methods.

Structs

CloseLoan

Instruction.

Collect

Instruction.

Configure

Instruction.

CreatePool

Instruction.

Deposit

Instruction.

MarginBorrow

Instruction.

MarginRefreshPosition

Instruction.

MarginRepay

Instruction.

RegisterLoan

Instruction.

Repay

Instruction.

Withdraw

Instruction.