Struct jet_margin_pool::MarginPool

#[repr(C, align(8))]
pub struct MarginPool {

    pub version: u8,
    pub pool_bump: [u8; 1],
    pub vault: Pubkey,
    pub fee_destination: Pubkey,
    pub deposit_note_mint: Pubkey,
    pub loan_note_mint: Pubkey,
    pub token_mint: Pubkey,
    pub token_price_oracle: Pubkey,
    pub address: Pubkey,
    pub config: MarginPoolConfig,
    pub borrowed_tokens: [u8; 24],
    pub uncollected_fees: [u8; 24],
    pub deposit_tokens: u64,
    pub deposit_notes: u64,
    pub loan_notes: u64,
    pub accrued_until: i64,

}

Account containing information about a margin pool, which services lending/borrowing operations.

Fields
version: u8pool_bump: [u8; 1]

The bump seed used to create the pool address

vault: Pubkey

The address of the vault account, which has custody of the pool’s tokens

fee_destination: Pubkey

The address of the account to deposit collected fees, represented as deposit notes

deposit_note_mint: Pubkey

The address of the mint for deposit notes

loan_note_mint: Pubkey

The address of the mint for the loan notes

token_mint: Pubkey

The token the pool allows lending and borrowing on

token_price_oracle: Pubkey

The address of the pyth oracle with price information for the token

address: Pubkey

The address of this pool

config: MarginPoolConfig

The configuration of the pool

borrowed_tokens: [u8; 24]

The total amount of tokens borrowed, that need to be repaid to the pool.

uncollected_fees: [u8; 24]

The total amount of tokens in the pool that’s reserved for collection as fees.

deposit_tokens: u64

The total amount of tokens available in the pool’s vault

deposit_notes: u64

The total amount of notes issued to depositors of tokens.

loan_notes: u64

The total amount of notes issued to borrowers of tokens

accrued_until: i64

The time the interest was last accrued up to

Implementations

impl MarginPool

pub fn signer_seeds(&self) -> Result<[&[u8]; 2]>

Get the seeds needed to sign for the vault

pub fn deposit(&mut self, amount: &FullAmount)

Record a deposit into the pool

pub fn withdraw(&mut self, amount: &FullAmount) -> Result<()>

Record a withdrawal from the pool

pub fn borrow(&mut self, amount: &FullAmount) -> Result<()>

Record a loan from the pool

pub fn repay(&mut self, amount: &FullAmount) -> Result<()>

Record a repayment of a loan

pub fn accrue_interest(&mut self, time: UnixTimestamp) -> bool

Accrue interest charges on outstanding borrows

Returns true if the interest was fully accumulated, false if it was only partially accumulated (due to significant time drift).

pub fn interest_rate(&self) -> Number

Gets the current interest rate for loans from this pool

pub fn utilization_rate(&self) -> Number

Gets the current utilization rate of the pool

pub fn collect_accrued_fees(&mut self) -> u64

Collect any fees accumulated from interest

Returns the number of notes to mint to represent the collected fees

pub fn calculate_prices(&self, pyth_price: &PriceFeed) -> Result<PriceResult>

Calculate the prices for the deposit and loan notes, based on the price of the underlying token.

pub fn calculate_full_amount(
&self,
current_notes: u64,
change: TokenChange,
action: PoolAction
) -> Result<FullAmount>

pub fn convert_amount(
&self,
amount: Amount,
action: PoolAction
) -> Result<FullAmount>

Convert the Amount to a FullAmount conisting of the appropriate proprtion of notes and tokens

pub fn total_uncollected_fees(&self) -> &Number

pub fn total_borrowed(&self) -> &Number